Move UnwrapExitCode to `errs`
