+++ internal/subshell/sscommon/sscommon.go
@@ -19,9 +19,6 @@ var (
 	// FailExecCmd represents a failure running a cmd
 	FailExecCmd = failures.Type("sscommon.fail.execcmd")
 
-	// FailExecCmdExit represents a cmd exit error failure
-	FailExecCmdExit = failures.Type("sscommon.fail.execcmdexit", failures.FailSilent)
-
 	// FailSignalCmd represents a failure sending a system signal to a cmd
 	FailSignalCmd = failures.Type("sscommon.fail.signalcmd")
 )
@@ -51,7 +48,7 @@ func Start(cmd *exec.Cmd) chan *failures.Failure {
 					return
 				}
 
-				fs <- FailExecCmdExit.Wrap(eerr)
+				fs <- failures.FailExecCmdExit.Wrap(eerr)
 				return
 			}
